requirements.txt
cohere==4.37
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
tqdm==4.66.2 